move write_aggregate into step file, and also extract write_repeat into separate function
